pyproject.toml
[project]
name = "stt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "torch>=2.7.0",
  "nemo_toolkit[asr]",
  "numpy<2",
  "cuda-python>=12.3",
  "sounddevice",
  "scipy",
  "PySide6",
  "webrtcvad",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

